An R6 class for Bayesian Ensemble

An R6::R6Class to interface with the Bayesian ensemble.

Public fields

stack

A data stack for the methods to have available for the use in the ensemble.

Methods

Public methods

ensemble$add()

ensemble$pop()

ensemble$run_ensemble()

ensemble$print()

ensemble$clone()

Method add()

Add a new method to the stack

Usage

ensemble$add(
  data,
  method,
  id_col,
  p_col,
  lfc_col,
  comp_col = "comparison",
  auxilary = "rest",
  do_rm = FALSE
)

Arguments

data

Input data with the results from a method.

method

The name to use to refer to the method in the stack.

id_col, p_col, lfc_col, comp_col

Name of the columns containing the unique identifier for each peptide, the p-values, log-fold change, and contast/comparison of the method, respectively.

auxilary

Additional paramters to store in the stack. Can be useful if do_rm is TRUE. Defaults to "rest" meaning the rest of the columns (other than the forementioned).

do_rm

Should the variable be removed from its current environment? Can be useful to reduce memory usage and avoid the need to rm after each method is added to the stack. ensemble will first try to find the environment of the data, if unsuccessful, it will try to remove it from the caller environment. Defaults to FALSE.

Method pop()

Remove a method from the stack and return it to stdout

Usage

ensemble$pop(method = length(self$stack))

Arguments

method

method index (integer) or name to return defaults to the most recently added method.

Method run_ensemble()

Run the ensemble

Usage

ensemble$run_ensemble(
  methods = "all",
  parallel_chains = 1,
  parallel_runs = 1,
  ...
)

Arguments

methods

Which methods to include in the ensemble.

parallel_chains

Number of parallel workers to use when running the sampling.

parallel_runs

The number of parallel comparison to run at the same time. Note that the total number of parallel instances will be parallel_chains\(\times\)parallel_runs. Can be efficient when there are many comparisons to run.

...

Additional parameters to rstan::sampling.

Returns

A tibble::tibble with the ensembled p-value. Column names will default to the first method in the stack.

Method print()

Printing method for the class.

Usage

ensemble$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

ensemble$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Setup variables
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Exemplify on the non-missing data
yeast <- tidyr::drop_na(yeast) %>%
   # Normalize and log-transform the data
   psrn("identifier")

# Fit the gamma regressions
gamma_reg_model <- yeast %>%
   calculate_mean_sd_trends(design) %>%
   fit_gamma_regression(formula = sd ~ mean)

limma_gr <- run_limma(
  yeast,
  design,
  contrast,
  gamma_reg_model,
  "identifier"
) %>%
   head(10)
limma_trend <- run_limma(
  yeast,
  design,
  contrast,
  NULL,
  "identifier"
) %>%
   dplyr::filter(identifier %in% limma_gr$identifier)

new_ensemble <- start_ensemble()
new_ensemble$add(limma_gr, "GR-limma", "identifier", "p_val", "lfc", do_rm = TRUE
)$add(
   limma_trend, "limma-trend", "identifier", "p_val", "lfc", do_rm = TRUE
)
ensemble_results <- new_ensemble$run_ensemble(iter = 150)
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 11
#> Chain 1:            adapt_window = 57
#> Chain 1:            term_buffer = 7
#> Chain 1: 
#> Chain 1: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 1: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 1: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 1: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 1: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 1: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 1: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 1: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 1: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 1: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 1: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 1: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.227 seconds (Warm-up)
#> Chain 1:                0.278 seconds (Sampling)
#> Chain 1:                0.505 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1.5e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: WARNING: There aren't enough warmup iterations to fit the
#> Chain 2:          three stages of adaptation as currently configured.
#> Chain 2:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 2:          the given number of warmup iterations:
#> Chain 2:            init_buffer = 11
#> Chain 2:            adapt_window = 57
#> Chain 2:            term_buffer = 7
#> Chain 2: 
#> Chain 2: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 2: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 2: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 2: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 2: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 2: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 2: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 2: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 2: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 2: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 2: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 2: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.162 seconds (Warm-up)
#> Chain 2:                0.163 seconds (Sampling)
#> Chain 2:                0.325 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 1.5e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: WARNING: There aren't enough warmup iterations to fit the
#> Chain 3:          three stages of adaptation as currently configured.
#> Chain 3:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 3:          the given number of warmup iterations:
#> Chain 3:            init_buffer = 11
#> Chain 3:            adapt_window = 57
#> Chain 3:            term_buffer = 7
#> Chain 3: 
#> Chain 3: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 3: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 3: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 3: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 3: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 3: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 3: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 3: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 3: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 3: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 3: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 3: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.13 seconds (Warm-up)
#> Chain 3:                0.16 seconds (Sampling)
#> Chain 3:                0.29 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 1.5e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: WARNING: There aren't enough warmup iterations to fit the
#> Chain 4:          three stages of adaptation as currently configured.
#> Chain 4:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 4:          the given number of warmup iterations:
#> Chain 4:            init_buffer = 11
#> Chain 4:            adapt_window = 57
#> Chain 4:            term_buffer = 7
#> Chain 4: 
#> Chain 4: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 4: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 4: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 4: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 4: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 4: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 4: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 4: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 4: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 4: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 4: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 4: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.134 seconds (Warm-up)
#> Chain 4:                0.113 seconds (Sampling)
#> Chain 4:                0.247 seconds (Total)
#> Chain 4: 
#> Warning: There were 10 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is 1.12, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
limma_gr <- new_ensemble$pop("limma_gr")