Run multiple imputation and limma

This function is an efficient wrapper that fits the needed gamma regressions, performs multiple imputation and testing with limma (see limmaUsersGuide). It is an efficient wrapper that generates need inputs for imputation and running run_limma with the possibility of using multidplyr-package to paralellize the computation. It also calls produces the mean-variance trends if plot_trend is TRUE.

multiple_imputation_and_limma(
  data,
  design,
  contrast_matrix,
  imputations,
  workers = 1,
  id_col = "id",
  .robust = TRUE,
  weights = TRUE,
  plot_trend = FALSE,
  formula_imputation = sd_p ~ mean,
  formula_weights = sd ~ mean,
  imp_pars = NULL,
  ...
)

Arguments

data

The data to run the pipeline on, missing values should have NA values.

design

A design matrix as produced by model.matrix.

contrast_matrix

A contrast matrix of comparisons to perfrom see makeContrasts for details.

imputations

Number of imputations to perfrome.

workers

Number of workers (processes) to run the pipeline with. Any value >1 will run the pipeline with parallel computing using the multidplyr-package.

id_col

A character for the name of the column containing the name of the features in data (e.g., peptides, proteins, etc.).

.robust

Input to limma robust option

weights

logical value if weights should be used.

plot_trend

Should the mean-variance trend with the gamma regression be plotted?

formula_imputation

Formula for the regression model used for imputation

formula_weights

Formula for the regression model used for weights

imp_pars

Imputation parameters to use; as inferred by get_imputation_pars.

...

Additional arguments to multi_trend_partitioning

Value

A tibble with each imputation as a row. The first column contains the imputation number, the second contains the imputed data, and the last column contains the results produced by run_limma.

Examples

# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important.
colnames(design) <- paste0("ng", c(50, 100))

# Generate the contrast matrix
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Normalize and log-transform the data
yeast <- psrn(yeast, "identifier")
# \donttest{
results <- multiple_imputation_and_limma(
  yeast,
  design,
  contrast,
  10, # Just running 10 imputation for the example purpose, but can be large > 1000
  1,  # Number of parallel workers to use (will drastically increase the speed)
  "identifier"
)
#> Estimating Imputation Paramters
#> Previous error: Inf 	>	Current error: 0.9295734 
#> Iteration time:
#>  7.172695 secs 
#> 
#> Previous error: 0.9295734 	>	Current error: 0.02261715 
#> Iteration time:
#>  7.52854 secs 
#> 
#> Previous error: 0.02261715 	>	Current error: 0.008648534 
#> Iteration time:
#>  7.308468 secs 
#> 
#> Previous error: 0.008648534 	>	Current error: 0.004244227 
#> Iteration time:
#>  7.287357 secs 
#> 
#> Previous error: 0.004244227 	>	Current error: 0.002974494 
#> Iteration time:
#>  7.354982 secs 
#> 
#> Previous error: 0.002974494 	>	Current error: 0.001724345 
#> Iteration time:
#>  7.307193 secs 
#> 
#> Previous error: 0.001724345 	<	Current error: 0.001800116 	Breaking 
#> Iteration time:
#>  7.322003 secs 
#> 
# }