Start a New Bayesian Ensemble

Creates a new ensemble instance.

start_ensemble()

Value

An ensemble object; see ensemble for details.

Examples

# Setup variables
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Exemplify on the non-missing data
yeast <- tidyr::drop_na(yeast) %>%
   # Normalize and log-transform the data
   psrn("identifier")

# Fit the gamma regressions
gamma_reg_model <- yeast %>%
   calculate_mean_sd_trends(design) %>%
   fit_gamma_regression(formula = sd ~ mean)

limma_gr <- run_limma(
  yeast,
  design,
  contrast,
  gamma_reg_model,
  "identifier"
) %>%
   head(10)
limma_trend <- run_limma(
  yeast,
  design,
  contrast,
  NULL,
  "identifier"
) %>%
   dplyr::filter(identifier %in% limma_gr$identifier)

new_ensemble <- start_ensemble()
new_ensemble$add(limma_gr, "GR-limma", "identifier", "p_val", "lfc", do_rm = TRUE
)$add(
   limma_trend, "limma-trend", "identifier", "p_val", "lfc", do_rm = TRUE
)
ensemble_results <- new_ensemble$run_ensemble(iter = 150)
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 11
#> Chain 1:            adapt_window = 57
#> Chain 1:            term_buffer = 7
#> Chain 1: 
#> Chain 1: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 1: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 1: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 1: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 1: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 1: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 1: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 1: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 1: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 1: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 1: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 1: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.125 seconds (Warm-up)
#> Chain 1:                0.16 seconds (Sampling)
#> Chain 1:                0.285 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1.5e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: WARNING: There aren't enough warmup iterations to fit the
#> Chain 2:          three stages of adaptation as currently configured.
#> Chain 2:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 2:          the given number of warmup iterations:
#> Chain 2:            init_buffer = 11
#> Chain 2:            adapt_window = 57
#> Chain 2:            term_buffer = 7
#> Chain 2: 
#> Chain 2: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 2: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 2: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 2: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 2: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 2: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 2: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 2: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 2: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 2: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 2: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 2: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.145 seconds (Warm-up)
#> Chain 2:                0.158 seconds (Sampling)
#> Chain 2:                0.303 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 1.5e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: WARNING: There aren't enough warmup iterations to fit the
#> Chain 3:          three stages of adaptation as currently configured.
#> Chain 3:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 3:          the given number of warmup iterations:
#> Chain 3:            init_buffer = 11
#> Chain 3:            adapt_window = 57
#> Chain 3:            term_buffer = 7
#> Chain 3: 
#> Chain 3: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 3: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 3: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 3: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 3: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 3: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 3: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 3: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 3: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 3: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 3: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 3: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.181 seconds (Warm-up)
#> Chain 3:                0.128 seconds (Sampling)
#> Chain 3:                0.309 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'mavis' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 1.5e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: WARNING: There aren't enough warmup iterations to fit the
#> Chain 4:          three stages of adaptation as currently configured.
#> Chain 4:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 4:          the given number of warmup iterations:
#> Chain 4:            init_buffer = 11
#> Chain 4:            adapt_window = 57
#> Chain 4:            term_buffer = 7
#> Chain 4: 
#> Chain 4: Iteration:   1 / 150 [  0%]  (Warmup)
#> Chain 4: Iteration:  15 / 150 [ 10%]  (Warmup)
#> Chain 4: Iteration:  30 / 150 [ 20%]  (Warmup)
#> Chain 4: Iteration:  45 / 150 [ 30%]  (Warmup)
#> Chain 4: Iteration:  60 / 150 [ 40%]  (Warmup)
#> Chain 4: Iteration:  75 / 150 [ 50%]  (Warmup)
#> Chain 4: Iteration:  76 / 150 [ 50%]  (Sampling)
#> Chain 4: Iteration:  90 / 150 [ 60%]  (Sampling)
#> Chain 4: Iteration: 105 / 150 [ 70%]  (Sampling)
#> Chain 4: Iteration: 120 / 150 [ 80%]  (Sampling)
#> Chain 4: Iteration: 135 / 150 [ 90%]  (Sampling)
#> Chain 4: Iteration: 150 / 150 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.152 seconds (Warm-up)
#> Chain 4:                0.145 seconds (Sampling)
#> Chain 4:                0.297 seconds (Total)
#> Chain 4: 
#> Warning: There were 21 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is 1.14, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
limma_gr <- new_ensemble$pop("limma_gr")