Normalization by Trimmed m Means

Method for trimmed m means normalization. It is based on the method described in Robinson and Oshlack (2010) . Though, instead of library size as was used in the original method, here we use the loading size which we define as the sum of all features. If no reference sample is specified, it uses the sample with the lowest coefficient of variation as default. All estimates are based on features without missing values.

tmm(
  data,
  trim_M = 0.3,
  trim_A = 0.05,
  log = TRUE,
  load_info = FALSE,
  target = NULL,
  reference_sample = NULL
)

Source

https://genomebiology.biomedcentral.com/articles/10.1186/gb-2010-11-3-r25

Arguments

data

data.frame containing the data to normalize

trim_M

percent of fold-change values to trim

trim_A

percent of means to trim

log

Return log2 transformed values?

load_info

Return loading info?

target

target columns to normalize, supports tidyselect-package syntax. By default, all numerical columns will be used in the normalization if not specified.

reference_sample

Specify a reference sample to normalize to, if not provided, the sample with the lowest coefficient of variation will be used

Value

data frame with normalized values if load_info=FALSE, if it is TRUE

then it returns a list with two tibbles. One tibble containing the normalized data and one containing the loading info as well as the estimated normalization factors.

References

Robinson MD, Oshlack A (2010). “A scaling normalization method for differential expression analysis of RNA-seq data.” Genome biology, 11(3), 1--9.

Examples

tmm(yeast)
#> # A tibble: 2,235 × 7
#>    identifier                       ng50_1 ng50_2 ng50_3 ng100_1 ng100_2 ng100_3
#>    <chr>                             <dbl>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#>  1 Cre01.g000450.t1.2|PACid:307886…   6.26   6.28   6.21    3.50    2.59   NA   
#>  2 Cre01.g000800.t1.2|PACid:307885…   7.94   7.98   7.89    7.82    7.84    7.86
#>  3 Cre01.g002050.t1.1|PACid:307891…   5.90   5.42   6.08    7.05    6.83    6.89
#>  4 Cre01.g002300.t1.2|PACid:307886…   6.49   6.22   5.73    6.60    6.51    6.27
#>  5 Cre01.g002300.t1.2|PACid:307886…   9.85  10.1    9.80   10.2    10.4    10.4 
#>  6 Cre01.g002300.t1.2|PACid:307886…   9.06   8.86   8.48   10.1    10.1    10.1 
#>  7 Cre01.g004350.t1.2|PACid:307886…   6.49   6.77   6.09    6.90    6.57    6.35
#>  8 Cre01.g004500.t1.2|PACid:307895…   1.82  NA     NA       6.63    6.80    6.35
#>  9 Cre01.g004500.t1.2|PACid:307895…  12.1   12.1   11.9    11.9    11.8    11.9 
#> 10 Cre01.g004500.t1.2|PACid:307895…   8.07   6.93   4.95    9.62    9.65    9.63
#> # ℹ 2,225 more rows