Spiked-in data set of reversibly oxidized cysteines

A dataset containing quantification of reversibly oxidized cysteines using Progenesis. True positives cysteines spiked-in from yeast at two different concentrations and true negatives from Chlamydomonas reinhardtii with the same concentration in all samples. To identify true positives one can use stringr::str_detect(yeast$identifier, 'YEAST'). For details see Berg et al. (2019) and if you use this dataset please cite the same paper.

yeast

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 2235 rows and 7 columns.

Source

https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-2619-6

References

Philip Berg, Evan W McConnell, Leslie M Hicks, Sorina C Popescu, George V Popescu (2019). “Evaluation of linear models and missing value imputation for the analysis of peptide-centric proteomics.” BMC bioinformatics, 20(2), 7--16.