An R6 class for Bayesian Ensemble

An R6::R6Class to interface with the Bayesian ensemble.

Public fields

stack

A data stack for the methods to have available for the use in the ensemble.

Methods

Public methods

ensemble$add()

ensemble$pop()

ensemble$run_ensemble()

ensemble$print()

ensemble$clone()

Method add()

Add a new method to the stack

Usage

ensemble$add(
  data,
  method,
  id_col,
  p_col,
  comp_col = "comparison",
  auxilary = "rest",
  do_rm = FALSE
)

Arguments

data

Input data with the results from a method.

method

The name to use to refer to the method in the stack.

id_col, p_col, lfc_col, comp_col

Name of the columns containing the unique identifier for each peptide, the p-values, log-fold change, and contast/comparison of the method, respectively.

auxilary

Additional paramters to store in the stack. Can be useful if do_rm is TRUE. Defaults to "rest" meaning the rest of the columns (other than the forementioned).

do_rm

Should the variable be removed from its current environment? Can be useful to reduce memory usage and avoid the need to rm after each method is added to the stack. ensemble will first try to find the environment of the data, if unsuccessful, it will try to remove it from the caller environment. Defaults to FALSE.

Method pop()

Remove a method from the stack and return it to stdout

Usage

ensemble$pop(method = length(self$stack))

Arguments

method

method index (integer) or name to return defaults to the most recently added method.

Method run_ensemble()

Run the ensemble

Usage

ensemble$run_ensemble(methods = "all", workers = 1)

Arguments

methods

Which methods to include in the ensemble.

workers

Number of parallel workers to use when running the ensemble. Can at most use one for each compario

...

Additional parameters to rstan::sampling.

Returns

A tibble::tibble with the ensembled p-value. Column names will default to the first method in the stack.

Method print()

Printing method for the class.

Usage

ensemble$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

ensemble$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Setup variables
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))
colnames(design) <- paste0("ng", c(50, 100))
contrast <- limma::makeContrasts(
  contrasts = "ng100-ng50",
  levels = design
)

# Exemplify on the non-missing data
yeast <- tidyr::drop_na(yeast) %>%
   # Normalize and log-transform the data
   psrn("identifier")

# Fit the gamma regressions
gamma_reg_model <- yeast %>%
   calculate_mean_sd_trends(design) %>%
   fit_gamma_regression(formula = sd ~ mean)

limma_gr <- run_limma(
  yeast,
  design,
  contrast,
  gamma_reg_model,
  "identifier"
) %>%
   head(10)
limma_trend <- run_limma(
  yeast,
  design,
  contrast,
  NULL,
  "identifier"
) %>%
   dplyr::filter(identifier %in% limma_gr$identifier)

new_ensemble <- start_ensemble()
new_ensemble$add(limma_gr, "GR-limma", "identifier", "p_val", "lfc", do_rm = TRUE
)$add(
   limma_trend, "limma-trend", "identifier", "p_val", "lfc", do_rm = TRUE
)
ensemble_results <- new_ensemble$run_ensemble()
#>  ..cutHeight not given, setting it to 91.3  ===>  99% of the (truncated) height range in dendro.
#>  ..done.
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `data = purrr::map2(data, w, private$ensemble)`.
#> Caused by warning:
#> ! There were 2 warnings in `dplyr::mutate()`.
#> The first warning was:
#> ℹ In argument: `w = purrr::map(tm, private$find_sd)`.
#> Caused by warning in `rgl.init()`:
#> ! RGL: unable to open X11 display
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
limma_gr <- new_ensemble$pop("limma_gr")